requirements.txt
alembic==0.9.3
backports.functools-lru-cache==1.5
blinker==1.4
click==6.7
cycler==0.10.0
dominate==2.3.1
itsdangerous==0.24
Jinja2==2.9.6
kiwisolver==1.0.1
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.2.3
mdsplus==7.7.12
numpy==1.15.4
pyparsing==2.3.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2018.9
six==1.10.0
SQLAlchemy==1.1.11
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
flask_bootstrap==3.3.7.1
flask_login==0.4.1
flask_mail==0.9.1
flask_migrate==2.0.4f
flask_moment==0.5.1
flask_script==2.0.6
flask_sqlalchemy==2.2
flask_wtf==0.14.2
